Cambios - Agregado clase Menu y refactorizacion
